Refactor CMS code
