+++ cms/api/base/blocks/map.py
@@ -1,0 +1,52 @@
+from decimal import Decimal
+from typing import Self
+from django.conf import settings
+import strawberry
+
+
+def generate_map_image(
+    latitude: Decimal, longitude: Decimal, width: int, height: int, zoom: int
+) -> str:
+    base = "https://api.mapbox.com/styles/v1/"
+    style = "mapbox/streets-v12"
+    token = f"access_token={settings.MAPBOX_PUBLIC_API_KEY}"
+
+    coordinates = f"{longitude},{latitude}"
+    size = f"{width}x{height}@2x"
+    marker = f"pin-s-heart+285A98({coordinates})"
+
+    return f"{base}{style}/static/{marker}/{coordinates},{zoom},0,13/{size}?{token}"
+
+
+@strawberry.type
+class CMSMap:
+    id: strawberry.ID
+    latitude: Decimal
+    longitude: Decimal
+    link: str
+    zoom: int
+
+    @classmethod
+    def from_block(cls, block) -> Self:
+        return cls(
+            id=block.id,
+            latitude=block.value["latitude"],
+            longitude=block.value["longitude"],
+            link=block.value["link"],
+            zoom=block.value["zoom"],
+        )
+
+    @strawberry.field
+    def image(
+        self,
+        info,
+        width: int = 1280,
+        height: int = 400,
+    ) -> str:
+        return generate_map_image(
+            latitude=self.latitude,
+            longitude=self.longitude,
+            width=width,
+            height=height,
+            zoom=self.zoom,
+        )
