+++ cms/api/schema.py
@@ -1,7 +1,7 @@
 from page.models import GenericPage as GenericPageModel
 
 import strawberry
-from api.types import GenericPage
+from api.page.types import GenericPage
 from wagtail.models import Site
