+++ cms/page/blocks/text_section.py
@@ -1,52 +1,6 @@
+from base.blocks.accordion import Accordion
+from base.blocks.cta import CTA
 from wagtail import blocks
-
-from wagtail.images.blocks import ImageChooserBlock
-
-
-class CTA(blocks.StructBlock):
-    label = blocks.CharBlock(required=False)
-    link = blocks.CharBlock(required=False)
-
-    class Meta:
-        label = "Call to Action"
-        icon = "site"
-
-
-class SimpleTextCard(blocks.StructBlock):
-    title = blocks.CharBlock()
-    body = blocks.RichTextBlock()
-    cta = CTA()
-
-    class Meta:
-        label = "Card: Simple Text"
-        icon = "doc-full"
-
-
-class SliderCardsSection(blocks.StructBlock):
-    cards = blocks.ListBlock(SimpleTextCard)
-
-    class Meta:
-        label = "Slider Cards Section"
-        icon = "crosshairs"
-
-
-class Accordion(blocks.StructBlock):
-    title = blocks.CharBlock()
-    body = blocks.RichTextBlock()
-    is_open = blocks.BooleanBlock(required=False, default=True)
-
-    class Meta:
-        icon = "arrow-down"
-
-
-class Map(blocks.StructBlock):
-    longitude = blocks.DecimalBlock(max_digits=9, decimal_places=6)
-    latitude = blocks.DecimalBlock(max_digits=9, decimal_places=6)
-    zoom = blocks.IntegerBlock(default=15)
-    link = blocks.URLBlock()
-
-    class Meta:
-        icon = "globe"
 
 
 class TextSection(blocks.StructBlock):
@@ -90,10 +44,3 @@ class TextSection(blocks.StructBlock):
 
     class Meta:
         icon = "doc-full"
-
-
-class BodyBlock(blocks.StreamBlock):
-    text_section = TextSection()
-    map = Map()
-    image = ImageChooserBlock()
-    slider_cards_section = SliderCardsSection()
